chore: collapsing out summaryhook subdirectory we don't really need
