--- ai_aside/summaryhook_aside/settings/common.py
@@ -1,12 +1,0 @@
-"""Default settings load from environment"""
-
-
-def plugin_settings(settings):
-    """
-    Get summary hook aside settings from calling application
-    """
-    env_tokens = getattr(settings, 'ENV_TOKENS', {})
-    settings.SUMMARY_HOOK_HOST = env_tokens.get('SUMMARY_HOOK_HOST', '')
-    settings.SUMMARY_HOOK_JS_PATH = env_tokens.get('SUMMARY_HOOK_JS_PATH', '')
-    settings.AISPOT_LMS_NAME = env_tokens.get('AISPOT_LMS_NAME', '')
-    settings.HTML_TAGS_TO_REMOVE = ['script', 'style']
