--- ai_aside/summaryhook_aside/settings/production.py
@@ -1,5 +1,0 @@
-"""Production settings values."""
-
-
-def plugin_settings(settings):  # pylint: disable=unused-argument
-    pass
