+++ ai_aside/text_utils.py
