--- tests/test_placeholder.py
@@ -1,12 +1,0 @@
-#!/usr/bin/env python
-"""
-Placeholder for tests for the `ai-aside` module.
-"""
-
-
-def test_placeholder():
-    """
-    Placeholder to allow pytest to succeed before real tests are in place.
-
-    (If there are no tests, it will exit with code 5.)
-    """
